mgr/dashboard: fix alert broken for multiple alerts
After carbon the alert panel was broken when there are multiple alerts
present (telemetry, motd, password expiration).

Applying carbon banner to the existing alert banner

Fixes: https://tracker.ceph.com/issues/66512
