+++ src/pybind/mgr/dashboard/frontend/src/app/core/layouts/workbench-layout/workbench-layout.component.scss
@@ -16,3 +16,8 @@
 .rgw-dashboard {
   background-color: vv.$body-bg-alt;
 }
+
+.cd-alert-container {
+  display: flex;
+  flex-direction: column;
+}
