+++ src/pybind/mgr/dashboard/frontend/src/app/core/navigation/navigation/navigation.component.html
@@ -1,10 +1,7 @@
 <div class="cd-navbar-main">
 <!-- ************************ -->
-<!-- NOTIFICATIONS/ALERTS     -->
+<!-- NOTIFICATIONS     -->
 <!-- ************************ -->
-<cd-pwd-expiration-notification></cd-pwd-expiration-notification>
-<cd-telemetry-notification></cd-telemetry-notification>
-<cd-motd></cd-motd>
 <cd-notifications-sidebar></cd-notifications-sidebar>
 <!-- ************************ -->
 <!-- HEADER                   -->
