+++ src/pybind/mgr/dashboard/frontend/src/app/shared/components/alert-panel/alert-panel.component.scss
@@ -10,3 +10,7 @@
 .alert {
   margin-bottom: 0;
 }
+
+cds-actionable-notification {
+  max-width: 100%;
+}
