+++ src/pybind/mgr/dashboard/frontend/src/app/shared/components/components.module.ts
@@ -15,7 +15,12 @@ import {
 import { ClickOutsideModule } from 'ng-click-outside';
 import { NgChartsModule } from 'ng2-charts';
 import { SimplebarAngularModule } from 'simplebar-angular';
-import { UIShellModule, ButtonModule, NotificationModule } from 'carbon-components-angular';
+import {
+  UIShellModule,
+  ButtonModule,
+  NotificationModule,
+  IconModule
+} from 'carbon-components-angular';
 
 import { MotdComponent } from '~/app/shared/components/motd/motd.component';
 import { DirectivesModule } from '../directives/directives.module';
@@ -80,7 +85,8 @@ import { UpgradableComponent } from './upgradable/upgradable.component';
     NgbTimepickerModule,
     UIShellModule,
     ButtonModule,
-    NotificationModule
+    NotificationModule,
+    IconModule
   ],
   declarations: [
     SparklineComponent,
