+++ src/pybind/mgr/dashboard/frontend/src/app/shared/components/telemetry-notification/telemetry-notification.component.html
@@ -1,5 +1,5 @@
 <cd-alert-panel *ngIf="displayNotification"
-                class="no-margin-bottom telemetry-notification"
+                class="no-margin-bottom"
                 [showTitle]="false"
                 size="slim"
                 [type]="notificationSeverity"
