+++ src/pybind/mgr/dashboard/frontend/src/app/shared/components/telemetry-notification/telemetry-notification.component.scss
@@ -21,11 +21,3 @@ a {
   color: darken(vv.$primary, 10);
   font-weight: bold;
 }
-
-.telemetry-notification {
-  bottom: 25px;
-  position: absolute;
-  right: 50%;
-  transform: translateX(50%);
-  z-index: 99999;
-}
